+++ config/routes.rb
@@ -1,6 +1,8 @@
 Rails.application.routes.draw do
   namespace :api do
-    resources :developer, only: [:show]
+    namespace :v1 do
+      resources :developer, only: [:show]
+    end
   end
   # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
 end
